wts_db/requirements.txt
PyYAML==3.12
SQLAlchemy==1.1.4
argparse==1.2.1
psycopg2==2.6.2
wsgiref==0.1.2
